Tibia 11.9x protocol (#8)
* Update server packet CyclopediaMapData

* Update server packet PlayerSkills

* Update server packet PreyRerollPrice

* Add example data for CyclopediaMapData packet changes

* Fix issue with parsing PlayerSkills packet.

* Add example data for CyclopediaMapData packet changes

* Add more example data for CyclopediaMapData changes

* Add missing CyclopediaMapData types
Fix exception if ReadString() has a length of 0

* Fix client MoveObject packet

* Start working on Stash related packets

* Add changes to PreyData server packet

* Add more packet changes

* Finish Stash packet

* Additional PreyData packet changes

* Fill out the CyclopediaMapData packet based on @TheSumm's work

* Rename PreyRerollPrice to PreyPrices as per client update

* Add new type (ActiveRaid) to CyclopediaMapData packet

* Fix misnamed values

* areaId should be ushort not uint

* Update unknown PreyPrices values

* Add unknown changes to StoreOffers packet

* Add known PreyData active changes

* Finish parsing PreyData packet

* More explicit TODOs around unknowns

* Update name for PreyPrices packet event

* Use corresponding login web service depending on client version

+++ TibiaAPI/Network/ClientPackets/RequestShopOffers.cs
@@ -10,7 +10,7 @@ namespace OXGaming.TibiaAPI.Network.ClientPackets
 
         public uint OfferId { get; set; }
 
-        public ushort Unknown { get; set; }
+        public uint Unknown { get; set; }
 
         public RequestShopOffers()
         {
@@ -34,7 +34,8 @@ namespace OXGaming.TibiaAPI.Network.ClientPackets
                 Category = message.ReadString();
             }
 
-            Unknown = message.ReadUInt16();
+            // TODO: Figure out this unknown.
+            Unknown = message.ReadUInt32();
             return true;
         }
